Move merge_binaries to Utils
